+++ ufotest/tests/mock.py
@@ -5,8 +5,13 @@ class MockTest(AbstractTest):
 
     name = "mock"
 
+    description = (
+        "This test case is simply a mock, which is used to test the ufotest software tools. The test case does "
+        "absolutely nothing and returns a simple MessageTestResult."
+    )
+
     def __init__(self, test_runner: TestRunner):
         AbstractTest.__init__(self, test_runner)
 
     def run(self):
-        return MessageTestResult(0, 'Hello World')
+        return MessageTestResult(0, 'Mock returned 0')